% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{condition}
\alias{condition}
\title{Constructor for condition}
\usage{
condition(subclass, message, call = sys.call(-1), ...)
}
\arguments{
\item{subclass}{a character name of the custom error class e.g. my_error}

\item{message}{a message to be associated with the custom error type}

\item{call}{system call that generated the error}
}
\value{
condition
}
\description{
Convenience function to construct condition
}
\author{
Adopted from Hadley Wickham
}
\keyword{internal}
