% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{custom_err}
\alias{custom_err}
\title{Generate (sub)class-specific error}
\usage{
custom_err(subclass, message, call = sys.call(-1), ...)
}
\arguments{
\item{subclass}{a character name of the custom error class e.g. my_error}

\item{message}{a message to be associated with the custom error type}

\item{call}{system call that generated the error}
}
\description{
This will allow creation of a custom error of type \verb{<subclass>}.
which inherits from type \code{error}. This enables setting up \code{tryCatch}
statements that could behave differently based on error type. (see
http://adv-r.had.co.nz/Exceptions-Debugging.html)
}
\author{
Adopted from Hadley Wickham
}
\keyword{internal}
