% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/d_test.R
\name{d_test}
\alias{d_test}
\title{generate two datasets for comparsion}
\usage{
d_test(n)
}
\arguments{
\item{n}{Number of observations for two datasets}
}
\value{
data 1 and data 2.
}
\description{
Generate two datasets for comparsion. The comparison are made on two main aspects : form and information
}
\examples{
#d<- d_test(n)
}
