% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiff_class.R
\name{ddiff_class}
\alias{ddiff_class}
\title{obtain types of variables from given two datasets}
\usage{
ddiff_class(dat1, dat2, key, ...)
}
\arguments{
\item{dat1}{First Data need to be compared.}

\item{dat2}{Second Data need to be compared.}

\item{key}{Unique identifiers to link two datasets}

\item{...}{potential argument to added}
}
\value{
The types of variables from given two datasets.
}
\description{
This function could obtain types of variables from given two datasets. Note that types of variables are defined as follows : continuous, categorical and binary
}
\examples{
d = test_data()
result <- ddiff_class(d$identical$new, d$identical$old, c("ID", "Days"))
data.tree::FromListSimple(ddiff_rpt)
str(ddiff_rpt)
}
