% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiff_info_multivariate.R
\name{ddiff_info_multivariate}
\alias{ddiff_info_multivariate}
\title{obtain and compare custom multivariate summary statistics from given two datasets}
\usage{
ddiff_info_multivariate(
  dat1,
  dat2,
  key,
  key_ex,
  measure_arg_con = list(cov = cov),
  measure_arg_cat = list(table = table),
  measure_arg_bin = list(table = table),
  dis_arg = list("euclidean", "manhattan", "gower"),
  ...
)
}
\arguments{
\item{dat1}{First Data need to be compared.}

\item{dat2}{Second Data need to be compared.}

\item{key}{Unique identifiers to link two datasets}

\item{key_ex}{variables that is excluded from variance covariance estimation}

\item{measure_arg_con}{list of function for summary statistics of continuous response. The defalut list is list(cov = cov)}

\item{measure_arg_cat}{list of function for summary statistics of categorical response}

\item{measure_arg_bin}{list of function for summary statistics of binary response}

\item{dis_arg}{list of distance for measuring the difference of variables in each datasets}

\item{...}{potential argument to added}
}
\value{
The sum of \code{x} and \code{y}.
#importFrom grDevices, col2rgb, dev.off, quartz, rainbow, rgb
#importFrom graphics, abline, contour, image, layout, legend, lines, matlines, matplot, par, persp, points, text, title
#importFrom stats, approxfun, cov, integrate, median, prcomp, rbinom, rnorm, runif, var
}
\description{
This function could obtain and compare custom multivariate summary statistics from given two datasets. Specially, the concurrent version involves obtaining the variance-covariance matrix of given two datasets
}
\examples{
d = test_data()
#result = ddiff_info_multivariate(d$identical$new, d$identical$old, , c("ID", "Days"), c("nEggs" , "nEggsRemain"))
#result$result_con
#result$result_cat
#result$result_bin
}
