% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiff_info_univariate.R
\name{ddiff_info_univariate}
\alias{ddiff_info_univariate}
\title{obtain and compare custom univariate summary statistics from given two datasets}
\usage{
ddiff_info_univariate(
  dat1,
  dat2,
  key,
  measure_arg_con = list(min = min, max = max, med = median, mean = mean, std =
    "standard error"),
  measure_arg_cat = list(table = table),
  measure_arg_bin = list(table = table),
  dis_arg = list("euclidean", "manhattan", "gower"),
  ...
)
}
\arguments{
\item{dat1}{First Data need to be compared.}

\item{dat2}{Second Data need to be compared.}

\item{key}{Unique identifiers to link two datasets}

\item{measure_arg_con}{list of function for summary statistics of continuous response. The defalut list is c("min", "max", "median", "mean", "standard error")}

\item{measure_arg_cat}{list of function for summary statistics of categorical response}

\item{measure_arg_bin}{list of function for summary statistics of binary response}

\item{dis_arg}{list of distance for measuring the difference of variables in each datasets}

\item{...}{potential argument to added}
}
\value{
TBC
}
\description{
This function could obtain and compare statistical information from given two datasets.
}
\examples{
d <- test_data()
ddiff_rpt <- ddiff_info_univariate(d$identical$new, d$identical$old, c("ID", "Days"))
ddiff_rpt
}
