% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{dfsig_get}
\alias{dfsig_get}
\title{Get cryptographic signature for a table}
\usage{
dfsig_get(df, parquet_v = "2.4")
}
\arguments{
\item{df}{a data.frame}

\item{pqrquet_v}{the version of parquet to be used for getting the signature}
}
\value{
a list with signature and parquet version used
}
\description{
Given a data.frame, it gets cryptographic signature for the
table when written on disk as a parquet file
}
