% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diff_attr.R
\name{diff_attr}
\alias{diff_attr}
\title{Attributes diff}
\usage{
diff_attr(df, df_ref)
}
\arguments{
\item{df}{a data.frame compared to the reference}

\item{df_ref}{reference data.frame relative to which comparison is made.}
}
\value{
a data.frame of logical with columns \code{colnames_match},
\code{rownames_match}, \code{attrnames_match}, and \code{other_attrs_match}
}
\description{
Compares attribues by checking colnames set and rownames set as
well as other attributes and whether they are all the identical
}
\examples{
diff_attr(cars,cars)
}
\keyword{internal}
