% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/diffable.R
\name{diffable}
\alias{diffable}
\title{diffable}
\usage{
diffable(df, df_ref, key = character(0), ...)
}
\arguments{
\item{df}{a data.frame compared to the reference}

\item{df_ref}{reference data.frame relative to which comparison is made.}

\item{key}{Unique identifiers to link two datasets}

\item{...}{potential argument to added}
}
\value{
The types of variables from given two datasets.
}
\description{
Creates a data object that valid for ddiff operation
}
\examples{
d = test_data()
ddiff_rpt <- diffable(d$identical$new, d$identical$old, c("ID", "Days"))
data.tree::FromListSimple(ddiff_rpt)
str(ddiff_rpt)

}
