% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fpca_multi_stat.R
\name{fpca_multi_stat}
\alias{fpca_multi_stat}
\title{obtain and compare custom summary statistics from given two datasets}
\usage{
fpca_multi_stat(dat1, dat2, key, ddiff_objective, ...)
}
\arguments{
\item{dat1}{First Data need to be compared.}

\item{dat2}{Second Data need to be compared.}

\item{key}{Unique identifiers to link two datasets}

\item{ddiff_objective}{First Data need to be compared.}

\item{...}{potential argument to added}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
obtain and compare custom summary statistics from given two datasets. The comparison are made on three main aspects : continuous, categorical and binary
}
\examples{
n = 100
d = generate_test_data(n)
#result = get_variable_stat(d$identical$new, d$identical$old, "id")
#result$result_con
#result$result_cat
#result$result_bin
}
