% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data.R
\name{generate_test_data}
\alias{generate_test_data}
\title{generate two datasets for comparsion}
\usage{
generate_test_data(n)
}
\arguments{
\item{n}{Number of observations for two datasets}
}
\value{
data 1 and data 2.
}
\description{
Generate two datasets for comparsion. The comparison are made on two main aspects : form and information
}
\examples{
n = 100
d = generate_test_data(n)
data.tree::FromListSimple(d)
}
