% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/generate_test_data2.R
\name{generate_test_data2}
\alias{generate_test_data2}
\title{generate ver1 two datasets for comparsion}
\usage{
generate_test_data2()
}
\arguments{
\item{n}{Number of observations for two datasets}
}
\value{
data 1 and data 2.
}
\description{
Generate two datasets for comparsion. The comparison are made on two main aspects : form and information
}
\examples{
d = generate_test_data2()
data.tree::FromListSimple(d)
}
