% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_class.R
\name{get_variable_class}
\alias{get_variable_class}
\title{obtain and compare types of variables from given two datasets}
\usage{
get_variable_class(dat1, dat2, key, ...)
}
\arguments{
\item{dat1}{First Data need to be compared.}

\item{dat2}{Second Data need to be compared.}

\item{key}{Unique identifiers to link two datasets}

\item{...}{potential argument to added}
}
\value{
The types of variables from given two datasets.
}
\description{
obtain and compare types of variables from given two datasets. The comparison are made on three main aspects : continuous, categorical and binary
}
\examples{
n = 100
d = generate_test_data(n)
result <- get_variable_class(d$identical$new, d$identical$old, "id")
result$result_con
result$result_cat
result$result_bin
#importFrom grDevices, col2rgb, dev.off, quartz, rainbow, rgb
#importFrom graphics, abline, contour, image, layout, legend, lines, matlines, matplot, par, persp, points, text, title
#importFrom stats, approxfun, cov, integrate, median, prcomp, rbinom, rnorm, runif, var
}
