% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_multi_stat.R
\name{get_variable_multi_stat}
\alias{get_variable_multi_stat}
\title{obtain and compare custom multivariate summary statistics from given two datasets}
\usage{
get_variable_multi_stat(
  dat1,
  dat2,
  key,
  measure_arg_con = list(cov = cov),
  measure_arg_cat = list(table = table),
  measure_arg_bin = list(table = table),
  ...
)
}
\arguments{
\item{dat1}{First Data need to be compared.}

\item{dat2}{Second Data need to be compared.}

\item{key}{Unique identifiers to link two datasets}

\item{measure_arg_con}{list of function for summary statistics of continuous response. The defalut list is list(cov = cov)}

\item{measure_arg_cat}{list of function for summary statistics of categorical response}

\item{measure_arg_bin}{list of function for summary statistics of binary response}

\item{...}{potential argument to added}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
obtain and compare custom multivariate summary statistics from given two datasets.
}
\examples{
n = 100
d = generate_test_data(n)
#result = get_variable_multi_stat(d$identical$new, d$identical$old, "id")
#result$result_con
#result$result_cat
#result$result_bin
}
