% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_variable_stat.R
\name{get_variable_stat}
\alias{get_variable_stat}
\title{obtain and compare custom summary statistics from given two datasets}
\usage{
get_variable_stat(
  dat1,
  dat2,
  key,
  measure_arg_con = list(min = min, max = max, med = median, mean = mean, std =
    "standard error"),
  measure_arg_cat = list(table = table),
  measure_arg_bin = list(table = table),
  ...
)
}
\arguments{
\item{dat1}{First Data need to be compared.}

\item{dat2}{Second Data need to be compared.}

\item{key}{Unique identifiers to link two datasets}

\item{measure_arg_con}{list of function for summary statistics of continuous response. The defalut list is c("min", "max", "median", "mean", "standard error")}

\item{measure_arg_cat}{list of function for summary statistics of categorical response}

\item{measure_arg_bin}{list of function for summary statistics of binary response}

\item{...}{potential argument to added}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
obtain and compare custom summary statistics from given two datasets. The comparison are made on three main aspects : continuous, categorical and binary
}
\examples{
n = 100
d = generate_test_data(n)
#result = get_variable_stat(d$identical$new, d$identical$old, "id")
#result$result_con
#result$result_cat
#result$result_bin
#importFrom grDevices, col2rgb, dev.off, quartz, rainbow, rgb
#importFrom graphics, abline, contour, image, layout, legend, lines, matlines, matplot, par, persp, points, text, title
#importFrom stats, approxfun, cov, integrate, median, prcomp, rbinom, rnorm, runif, var
}
