% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/guess_col_class.R
\name{guess_col_class}
\alias{guess_col_class}
\title{Guess column class}
\usage{
guess_col_class(
  df,
  guess_fun = function(x) readr::guess_parser(as.character(x))
)
}
\arguments{
\item{a}{data.frame}
}
\description{
Given a data.frame it guess its column classes
}
\keyword{internal}
