% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{is.condition}
\alias{is.condition}
\title{Assert for condition}
\usage{
is.condition(x)
}
\arguments{
\item{x}{Condition}
}
\value{
TRUE or FALSE
}
\description{
Asserts that x is a condition
}
\author{
Adopted from Hadley Wickham
}
\keyword{internal}
