% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_ddiff_input.R
\name{make_ddiff_input}
\alias{make_ddiff_input}
\title{format longitudinal dataset for functional data analysis}
\usage{
make_ddiff_input(
  IDs = NULL,
  tVec,
  yVec,
  na.rm = FALSE,
  sort = FALSE,
  deduplicate = FALSE
)
}
\arguments{
\item{IDs}{IDs}

\item{tVec}{tVec}

\item{yVec}{yVec}

\item{na.rm}{remove NA T/F}

\item{sort}{to sort or not T/F}

\item{deduplicate}{to dedup or not T/F}
}
\value{
data 1 and data 2.
}
\description{
This function could form dataset for ddiff and analysis
}
\examples{
d = test_data()
data.tree::FromListSimple(d)
}
