% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/match_record.R
\name{match_record}
\alias{match_record}
\title{Match records across two data.frames}
\usage{
match_record(diffable_obj, mincol_match_rate = 0.5)
}
\arguments{
\item{df}{}
}
\description{
Match records across two data.frames
}
