% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pca_multi_stat.R
\name{pca_multi_stat}
\alias{pca_multi_stat}
\title{obtain and compare custom summary statistics from given two datasets}
\usage{
pca_multi_stat(ddiff_objective, ...)
}
\arguments{
\item{...}{potential argument to added}

\item{dat1}{First Data need to be compared.}

\item{dat2}{Second Data need to be compared.}

\item{key}{Unique identifiers to link two datasets}

\item{measure_arg_con}{list of function for summary statistics of continuous response. The defalut list is c("min", "max", "median", "mean", "standard error")}

\item{measure_arg_cat}{list of function for summary statistics of categorical response}

\item{measure_arg_bin}{list of function for summary statistics of binary response}
}
\value{
The sum of \code{x} and \code{y}.
}
\description{
obtain and compare custom summary statistics from given two datasets. The comparison are made on three main aspects : continuous, categorical and binary
}
\examples{
n = 100
d = generate_test_data(n)
#result = get_variable_stat(d$identical$new, d$identical$old, "id")
#result$result_con
#result$result_cat
#result$result_bin
}
