% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ddiff_form.R
\name{ddiff_form}
\alias{ddiff_form}
\title{compare given two datasets in form level}
\usage{
ddiff_form(dat1, dat2, key, ...)
}
\arguments{
\item{dat1}{First Data need to be compared.}

\item{dat2}{Second Data need to be compared.}

\item{key}{Unique identifiers to link two datasets}

\item{...}{potential argument to added}
}
\value{
ddiff which contains indicator of above aspects. Also, the last element include existing comparedf object.
}
\description{
This function could obtain and compare empirical information of row, column and attributes from given two datasets. The comparison are made on following aspects :
\enumerate{
\item whether the types of variables from given two datasets are the same
\item whether column names from given two datasets are the same
\item whether row names from given two datasets are the same
\item whether attributes from given two datasets are the same
\item whether the information given two datasets are the same (shuffled rows or columns)
\item whether the given two datasets are equivalent in R memory
}
}
\examples{
d <- test_data()
ddiff_rpt <- ddiff_form(d$identical$new, d$identical$old, c("ID", "Days"))
ddiff_rpt
}
